requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
async-lru==2.0.4
attrs==24.3.0
beautifulsoup4==4.12.3
black==24.10.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
colorlog==6.9.0
cryptography==42.0.8
frozenlist==1.5.0
greenlet==3.1.1
idna==3.7
instagrapi==2.1.3
isort==5.13.2
lxml==5.3.0
miniupnpc==2.2.8
multidict==6.1.0
mypy==1.14.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
playwright==1.49.1
praw==7.8.1
prawcore==2.4.0
propcache==0.2.1
pycparser==2.22
pycryptodomex==3.21.0
pycurl==7.45.4
pydantic==2.10.1
pydantic_core==2.27.1
pyee==12.0.0
PySocks==1.7.1
requests==2.32.3
requests-oauthlib==1.3.1
ruff==0.9.1
setuptools==71.0.4
soupsieve==2.6
termcolor==2.4.0
tweepy==4.14.0
types-pycurl==7.45.4.20241216
types-requests==2.32.0.20240712
typing_extensions==4.12.2
update-checker==0.18.0
UPnPy==1.1.8
urllib3==2.2.2
websocket-client==1.8.0
yarl==1.18.3
